rename metrics_no_db to metrics_redis
