+++ awx/main/analytics/metrics_redis.py
