+++ awx/main/queue.py
@@ -8,7 +8,7 @@ import redis
 
 # Django
 from django.conf import settings
-import awx.main.analytics.metrics_no_db as metrics_no_db
+import awx.main.analytics.metrics_redis as metrics_redis
 
 __all__ = ['CallbackQueueDispatcher']
 
@@ -34,4 +34,4 @@ class CallbackQueueDispatcher(object):
 
     def dispatch(self, obj):
         self.connection.rpush(self.queue, json.dumps(obj, cls=AnsibleJSONEncoder))
-        metrics_no_db.hincrby('callback_receiver_events_insert_redis', 1)
+        metrics_redis.hincrby('callback_receiver_events_insert_redis', 1)
