+++ awx/main/wsbroadcast.py
@@ -141,8 +141,8 @@ class BroadcastWebsocketTask(WebsocketTask):
                     continue
                 (group, message) = unwrap_broadcast_msg(payload)
                 if group == "metrics":
-                    import awx.main.analytics.metrics_no_db as metrics_no_db
-                    metrics_no_db.store_metrics(message)
+                    import awx.main.analytics.metrics_redis as metrics_redis
+                    metrics_redis.store_metrics(message)
                 await self.channel_layer.group_send(group, {"type": "internal.message", "text": message})
